+++ data/mypkg/utils.py
@@ -1,9 +1,0 @@
-# utils.py - 提供工具函数
-
-def greet(name):
-    """一个简单的问候函数"""
-    return f"你好, {name}!"
-
-def add(a, b):
-    """加法"""
-    return a + b